Add Java Client interface for spring projects
This commit showcases the Java Interface client support in Spring
Framework 6. A SpringProjectsClient interface defines the access
that we need, using a subset of the API that can be found on the
server-side.

A proxy can be created from the interface using a few lines of code,
as demonstrated in the main app.